Move `DbOp` enum from `models/datastore` to `models/sql` and use in `db_operations` module
Better reflects the scope of the enum
