+++ src/modules/db_operations/main.py
@@ -21,12 +21,14 @@ from caendr.services.cloud.postgresql import get_db_conn_uri, get_db_timeout, db
 from caendr.services.cloud.secret import get_secret
 from operations import execute_operation
 
+from caendr.models.sql import DbOp
+
 
 # Load environment variables
 MODULE_DB_OPERATIONS_BUCKET_NAME = get_env_var('MODULE_DB_OPERATIONS_BUCKET_NAME')
 ETL_LOGS_BUCKET_NAME             = get_env_var('ETL_LOGS_BUCKET_NAME')
 EXTERNAL_DB_BACKUP_PATH          = get_env_var('EXTERNAL_DB_BACKUP_PATH')
-DB_OP                            = get_env_var('DATABASE_OPERATION')
+DATABASE_OPERATION               = get_env_var('DATABASE_OPERATION')
 EMAIL                            = get_env_var('EMAIL',        can_be_none=True)
 OPERATION_ID                     = get_env_var('OPERATION_ID', can_be_none=True)
 
@@ -105,6 +107,12 @@ def run():
   use_mock_data = get_env_var('USE_MOCK_DATA', False, var_type=bool)
   reload_files  = get_env_var('RELOAD_FILES',  True,  var_type=bool)
 
+  # Parse database operation
+  try:
+    db_op = DbOp[DATABASE_OPERATION]
+  except:
+    logger.error(f'Unknown database operation {DATABASE_OPERATION}')
+
   # Parse species list
   species = parse_species_list( get_env_var('SPECIES_LIST', can_be_none=True) )
   species_string = '[' + ', '.join(species) + ']' if species else 'all'
@@ -112,16 +120,16 @@ def run():
   text = ""
 
   try:
-    execute_operation(app, db, DB_OP, species=species, reload_files=reload_files)
+    execute_operation(app, db, db_op, species=species, reload_files=reload_files)
     text = text + f"\n\nStatus: OK"
-    text = text + f"\nOperation: {DB_OP}"
+    text = text + f"\nOperation: {db_op.name}"
     text = text + f"\nOperation ID: {OPERATION_ID}"
     text = text + f"\nEnvironment: { get_env_var('ENV', 'n/a') }"
     text = text + f"\nSpecies: {species_string}"
 
   except Exception as e:
     text = text + f"\nStatus: ERROR"
-    text = text + f"\nOperation: {DB_OP}"
+    text = text + f"\nOperation: {db_op.name}"
     text = text + f"\nOperation ID: {OPERATION_ID}"
     text = text + f"\nEnvironment: { get_env_var('ENV', 'n/a') }"
     text = text + f"\nSpecies: {species_string}"
@@ -149,7 +157,7 @@ def run():
     send_email({
       "from": f'CaeNDR <{NO_REPLY_EMAIL}>',
       "to": EMAIL,
-      "subject": f"ETL finished for operation: {DB_OP} in {elapsed} seconds",
+      "subject": f"ETL finished for operation: {db_op.name} in {elapsed} seconds",
       "text": text,
     })
